feat(headless): export analytics actions for all features
https://coveord.atlassian.net/browse/KIT-277
