+++ packages/headless/src/features/facets/facet-set/facet-set-analytics-actions-utils.ts
@@ -1,0 +1,23 @@
+import {
+  CategoryFacetSection,
+  DateFacetSection,
+  FacetSection,
+  NumericFacetSection,
+} from '../../../state/state-sections';
+import {RangeFacetSortCriterion} from '../range-facets/generic/interfaces/request';
+import {FacetSortCriterion} from './interfaces/request';
+
+export type SectionNeededForFacetMetadata = FacetSection &
+  CategoryFacetSection &
+  DateFacetSection &
+  NumericFacetSection;
+
+export type FacetUpdateSortMetadata = {
+  facetId: string;
+  criterion: FacetSortCriterion | RangeFacetSortCriterion;
+};
+
+export type FacetSelectionChangeMetadata = {
+  facetId: string;
+  facetValue: string;
+};
