+++ packages/headless/src/features/facets/range-facets/date-facet-set/date-facet-set-slice.test.ts
@@ -13,7 +13,7 @@ import * as RangeFacetReducers from '../generic/range-facet-reducers';
 import * as FacetReducers from '../../generic/facet-reducer-helpers';
 import {executeSearch} from '../../../search/search-actions';
 import {buildMockSearch} from '../../../../test/mock-search';
-import {logGenericSearchEvent} from '../../../analytics/analytics-actions';
+import {logSearchEvent} from '../../../analytics/analytics-actions';
 import {buildMockDateFacetValue} from '../../../../test/mock-date-facet-value';
 import {
   DateFacetSetState,
@@ -113,7 +113,7 @@ describe('date-facet-set slice', () => {
     const search = buildMockSearch();
     dateFacetSetReducer(
       state,
-      executeSearch.fulfilled(search, '', logGenericSearchEvent({evt: 'foo'}))
+      executeSearch.fulfilled(search, '', logSearchEvent({evt: 'foo'}))
     );
 
     expect(
