+++ packages/headless/src/features/product-recommendations/product-recommendations-analytics.actions.ts
@@ -1,11 +1,8 @@
 import {
-  AnalyticsType,
-  makeAnalyticsAction,
-} from '../analytics/analytics-actions';
-import {
   ProductRecommendationAnalyticsProvider,
   StateNeededByProductRecommendationsAnalyticsProvider,
 } from '../../api/analytics/product-recommendations-analytics';
+import {AnalyticsType, makeAnalyticsAction} from '../analytics/analytics-utils';
 
 /**
  * Logs a search event with an `actionCause` value of `recommendationInterfaceLoad`.
