+++ packages/headless/src/features/recommendation/recommendation-analytics-actions.ts
@@ -1,7 +1,4 @@
-import {
-  AnalyticsType,
-  makeAnalyticsAction,
-} from '../analytics/analytics-actions';
+import {AnalyticsType, makeAnalyticsAction} from '../analytics/analytics-utils';
 
 /**
  * Logs a search event with an `actionCause` value of `recommendationInterfaceLoad`.
