+++ packages/headless/src/features/sort-criteria/sort-criteria-analytics-actions.ts
@@ -1,7 +1,4 @@
-import {
-  makeAnalyticsAction,
-  AnalyticsType,
-} from '../analytics/analytics-actions';
+import {AnalyticsType, makeAnalyticsAction} from '../analytics/analytics-utils';
 import {getSortCriteriaInitialState} from './sort-criteria-state';
 
 /**
